increase lscratch size

+++ construct_lfp_bash_scripts.py
@@ -184,9 +184,11 @@ def write_session_scripts(subj_path, sess, nsp_suffix, session_nsx_fpath, sessio
 	if os.path.isdir(session_dir) is False:
 		os.mkdir(session_dir)
 
-	for log_file in glob.glob(session_dir + "/*.log"):
+	log_glob = glob.glob(session_dir + "/*.log")
+	if log_glob != []:
 		print(" ... removing old log files", end="")
-		os.remove(log_file)
+		for log_file in log_glob:
+			os.remove(log_file)
 
 	time_log_fpath = session_dir + "/time.log"
 	lfp_sbatch_file = open(session_dir + "/" + bash_fname, 'w')
